Move metadata_jax.py to sampling_metadata.py
